took out the experimental block transfer stuff
git-svn-id: svn://orca.physics.unc.edu/Orca@348 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/Camac/DGF4c/ORDGF4cModel.h
@@ -89,7 +89,7 @@ union {
 		short decimation;
 		long numOscPoints;
 		unsigned short oscModeData[4][8192];
-		unsigned short oscEnabledMask;
+		int oscEnabledMask;
 		float energyRiseTime[4];
 		float energyFlatTop[4];
 		float triggerRiseTime[4];
